move ChatHeader to tailwind
