+++ chat-react2/src/components/Home/Main/Main.tsx
@@ -4,7 +4,7 @@ import { useChat } from "@connectycube/use-chat";
 import { IoMdAttach } from "react-icons/io";
 import { IoIosArrowDown } from "react-icons/io";
 import Message from "./Message/Message";
-import ChatHeader from "./ChatHeader/ChatHeader";
+import ChatHeader from "./ChatHeader";
 import ChatInfo from "./ChatInfo/ChatInfo";
 import "./Main.scss";
